improve seen hooks structure
