+++ client/reader/data/seen-posts/index.ts
@@ -1,3 +1,5 @@
 export * from './use-is-seen-enabled';
+export * from './use-can-mark-seen';
+export * from './use-is-seen-visible';
 export * from './use-mutations';
 export * from './with-seen-posts-mutations';
