+++ client/reader/data/seen-posts/use-is-seen-visible.ts
@@ -1,0 +1,22 @@
+import { useSiteSubscriptionOrganizationId } from 'calypso/reader/data/site-subscriptions';
+import { SeenArgs, useIsSeenEnabled, isPostAnAFKPost } from './use-is-seen-enabled';
+
+/**
+ * Returns true if the user can apply the seen state to a post, false otherwise.
+ *
+ * Mainly we need this because we want to show the seen state for AFK posts, but not allow users to mark them as seen.
+ */
+export function useIsSeenVisible( { feedId, blogId, post }: SeenArgs ): boolean {
+	const isSeenEnabled = useIsSeenEnabled( { feedId, blogId } );
+	const organizationId = useSiteSubscriptionOrganizationId( feedId, blogId );
+
+	if ( ! isSeenEnabled ) {
+		return false;
+	}
+
+	if ( isPostAnAFKPost( organizationId, post ) ) {
+		return true;
+	}
+
+	return Boolean( post?.is_seen );
+}
